+++ fortepyan/audio/render.py
@@ -35,14 +35,16 @@ def midi_to_wav(midi: Union[structures.MidiFile, structures.MidiPiece], wavpath:
     """
     if isinstance(midi, structures.MidiPiece):
         midi = midi.to_midi()
+
     # This will be deleted
     tmp_midi_path = tempfile.mkstemp(suffix=".mid")[1]
 
     # Add an silent event to make sure the final notes
     # have time to ring out
-    end_time = midi.get_end_time() + 0.2
+    end_time = midi.duration + 0.2
     pedal_off = pretty_midi.ControlChange(64, 0, end_time)
-    midi.instruments[0].control_changes.append(pedal_off)
+
+    midi._midi.instruments[0].control_changes.append(pedal_off)
 
     midi.write(tmp_midi_path)