+++ fortepyan/audio/soundfont.py
@@ -1,4 +1,4 @@
-import urllib
+import urllib.request
 from pathlib import Path
 
 import appdirs